LIBFCREPO-797. Added ZipFile class.
* during import, choose binary source type based on binaries location URI scheme
* refactored binary sources and subclasses into plastron.files module

https://issues.umd.edu/browse/LIBFCREPO-797
